old/requirements.txt
discord.py == 2.2.2
py-cord == 2.4.0
python-dotenv
PyYAML
requests
typing
youtube-dl == 2021.12.17
pytube == 12.1.3